+++ src/main/java/com/driver/services/ImageService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ImageService {
@@ -17,29 +18,40 @@ public class ImageService {
 
     public Image addImage(Integer blogId, String description, String dimensions){
         //add an image to the blog
-        Blog blog=blogRepository2.findById(blogId).get();
-        Image image=new Image();
-        image.setBlog(blog);
-        image.setDescription(description);
-        image.setDimensions(dimensions);
-        imageRepository2.save(image);
-        return image;
+        Optional<Blog> optionalBlog=blogRepository2.findById(blogId);
+        if(optionalBlog.isPresent()) {
+            Blog blog=optionalBlog.get();
+            Image image = new Image();
+            image.setBlog(blog);
+            image.setDescription(description);
+            image.setDimensions(dimensions);
+            imageRepository2.save(image);
+            return image;
+        }
+        return null;
     }
 
     public void deleteImage(Integer id){
-        imageRepository2.deleteById(id);
+        Optional<Image> optionalImage=imageRepository2.findById(id);
+        if(optionalImage.isPresent()) {
+            imageRepository2.deleteById(id);
+        }
     }
 
     public int countImagesInScreen(Integer id, String screenDimensions) {
         //Find the number of images of given dimensions that can fit in a screen having `screenDimensions`
-        Image image=imageRepository2.findById(id).get();
-        String[] dim=image.getDimensions().split("X");
-        int inLen=Integer.parseInt(dim[0]);
-        int inWid=Integer.parseInt(dim[1]);
+        Optional<Image> optionalImage=imageRepository2.findById(id);
+        if(optionalImage.isPresent()) {
+            Image image=optionalImage.get();
+            String[] dim = image.getDimensions().split("X");
+            int inLen = Integer.parseInt(dim[0]);
+            int inWid = Integer.parseInt(dim[1]);
 
-        String[] dim2=screenDimensions.split("X");
-        int outLen=Integer.parseInt(dim2[0]);
-        int outWid=Integer.parseInt(dim2[1]);
-        return (outLen/inLen)*(outWid/inWid);
+            String[] dim2 = screenDimensions.split("X");
+            int outLen = Integer.parseInt(dim2[0]);
+            int outWid = Integer.parseInt(dim2[1]);
+            return (outLen / inLen) * (outWid / inWid);
+        }
+        return 0;
     }
 }
